+++ src/index.html
@@ -76,7 +76,7 @@
     <img src="" alt="">
     <img src="" alt="">
   </footer>
-  <script type="module" src="main.js"></script>
+  <script type="module" src="data.js"></script>
 </body>
 
 </html>
